Rename tslib directory to lib
